% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data4consumerscience-package.R
\docType{package}
\name{data4consumerscience-package}
\alias{data4consumerscience}
\alias{data4consumerscience-package}
\title{data4consumerscience: A collection of food consumer datasets}
\description{
This package can be used for teaching data analytics for food consumer science.
}
\author{
\strong{Maintainer}: Morten Arendt Rasmussen \email{mortenr@food.ku.dk} (\href{https://orcid.org/0000-0001-7431-5206}{ORCID})

}
\keyword{internal}
