% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pork.R
\docType{data}
\name{pork}
\alias{pork}
\title{Data of pork consumption}
\format{
A tibble with 1633 rows and 58 variables:
\describe{
\item{id}{dbl person id}
\item{COUNTRY}{COUNTRY}
\item{Locality}{residency rual / urban}
}
}
\source{
\url{https://doi.org/10.1016/j.meatsci.2011.02.016}
}
\usage{
pork
}
\description{
From: Verbeke, W., Pérez-Cueto, F. J. A., Grunert, K. G. (2011). To eat or not to eat pork, how frequently and how varied? Insights from the quantitative Q-PorkChains consumer survey in four European countries. Meat Science, 88, 619-626.
}
\keyword{datasets}
